ASSESMENTS/5- python advance/P_HOTEL_MANAGEMENT_SYSTEM_/requirement.txt
mysql-connector-python==8.2.0
pillow==10.2.0
protobuf==4.21.12
